+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v3.0.115(DEC 24, 2019)
+
+- Added `reactions` in `BaseMessage`.
+  - Added `addReaction()`, `deleteReaction()` in `BaseChannel`.
+  - Added `onReactionUpdated()` in `ChannelHandler`.
+  - Added `applyReactionEvent()` in `BaseMessage`.
+  - Added `includeReaction` parameter to `get**MessagesByID()`, `get**MessagesByTimestamp()` in `BaseChannel`.
+  - Added `includeReaction` parameter to `getMessageChangeLogsByID()`, `getMessageChangeLogsByTimestamp()` in `BaseChannel`.
+- Added always push notification support.
+- Added `markAsDelivered()`, `getDeliveryReceipt()` in `GroupChannel`.
+  - Added `onDeliveryReceiptUpdated` in `ChannelHandler`.
+- Improved stability.
+
 ## v3.0.114(DEC 18, 2019)
 
 - Bug-fix in file uploading.